requirements.txt
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0
enum34==1.1.6
Flask==0.12.2
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pkg-resources==0.0.0
pycparser==2.18
pymediainfo==2.1.9
pyOpenSSL==17.2.0
python-dateutil==2.6.1
requests==2.18.2
six==1.10.0
tmdbsimple==1.7.0
urllib3==1.22
Werkzeug==0.12.2
